Release 1.0.9 improvements -- II (#27)
* WIP:
1. Don't clobber discription
2. Add Obsidian URL to tile on ticktick side only
3. Reset file lock on exception

* WIP Fixes
TickTick Task description will no longer be over-written
Obsidian URL added to Task Title
Accomodate ALL versions of Task markdown.
Bug Fix: ALL TickTick side modifications will sync.

* Update Readme with deleted files warning.

* Prevent Task Description over-write. Handle Items
1. TickTick Task description will no longer be over-written
2. Obsidian URL added to Task Title instead of over-writing description
3. Bi-Directional sync of Task Items
4. Accomodate ALL versions of Task markdown.
5. Bug Fix: ALL TickTick side modifications will sync.
6. Only allow a Project to be default for one single file.
   (limitation of current task sync)

* 1.0.8 update

* Refactor and Take care of issues
Fix: https://github.com/thesamim/TickTickSync/issues/10
Add Warnings on Task Deletion:
https://github.com/thesamim/TickTickSync/issues/6

* Complete Release 1.0.8 improvements
Items and Contents now correctly handled
Time Zone deltas should be resolved
External Task editing is handled

* Complete Release 1.0.8 improvements
Clean up overzealous logging
Accomodate New World Order Logon
Fix up Read me file.

* Complete Release 1.0.8 improvements
Fix login flow issues.
Pass token to ticktick api wrapper

* Complete Release 1.0.8 improvements
Update Change Log

* Complete Release 1.0.8 improvements
Fix File Name Case

* Release 1.0.9 improvements
Fix typo in README.md
Add deletion reason to confirmation dialog
Add deletion confirmation to some edge cases.

* Release 1.0.9 improvements
OverAgressive confirmation request.

* Release 1.0.9 improvements
Fix OverAgressive confirmation request.

+++ src/syncModule.ts
@@ -862,15 +862,19 @@ export class SyncMan {
 			const reallyDeletedTickTickTasks = deletedTickTickTasks.filter(task => deletedTasks.some(t => t.taskId === task.id));
 			// this.dumpArray('== reallyDeletedTickTickTasks deleted from TickTick:', reallyDeletedTickTickTasks);
 
-			const taskTitlesForConfirmation = reallyDeletedTickTickTasks.map((task: ITask) => task.id)
 
-			const bConfirm = await  this.confirmDeletion(taskTitlesForConfirmation, "tasks deleted from TickTick");
+			if (reallyDeletedTickTickTasks.length > 0) {
+				const taskTitlesForConfirmation = reallyDeletedTickTickTasks.map((task: ITask) => task.id)
 
-			if (bConfirm) {
-				for (const task of reallyDeletedTickTickTasks) {
-					await this.plugin.fileOperation?.deleteTaskFromFile(task);
-					await this.plugin.cacheOperation?.deleteTaskFromCache(task.id)
-					bModifiedFileSystem = true;
+				const bConfirm = await this.confirmDeletion(taskTitlesForConfirmation, "tasks deleted from TickTick");
+
+				if (bConfirm) {
+					for (const task of reallyDeletedTickTickTasks) {
+						await this.plugin.fileOperation?.deleteTaskFromFile(task);
+						await this.plugin.cacheOperation?.deleteTaskFromCache(task.id)
+						bModifiedFileSystem = true;
+					}
+
 				}
 			}
